Move setup to BeforeTemplateRendered event listener.
The point is that this is an interactive app that need not be fully
setup for non-interactive request. OTOH, we need the app to be "booted"
in order to intercept the app-menu click. Perhaps it would be good to
find out if we can move it out of the way for all non interactive
requests, but ATM I just do not know how to do this.